# swagger_output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Aeko Enhanced API",
    "version": "2.0.0",
...
    "contact": {
      "name": "Aeko Development Team",
      "email": "dev@aeko.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development Server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Authorization header using the Bearer scheme. Example: 'Authorization: Bearer {token}'"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/chat": {
      "post": {
        "summary": "Interact with the Smart Bot",
        "tags": [
          "Bot"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "message": {
                    "type": "string"
                  }
                },
                "required": [
                  "message"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Bot reply returned successfully"
          },
          "400": {
            "description": "Bad request"
          },
          "403": {
            "description": "Bot is disabled"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/chat/send-message": {
      "post": {
        "summary": "Send a message",
        "tags": [
          "Chat"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "recipientId",
                  "message"
                ],
                "properties": {
                  "recipientId": {
                    "type": "string"
                  },
                  "message": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Message sent successfully"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/chat/bot-settings": {
      "put": {
        "summary": "Update Smart Bot settings",
        "tags": [
          "Bot"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "botEnabled",
                  "botPersonality"
                ],
                "properties": {
                  "botEnabled": {
                    "type": "boolean"
                  },
                  "botPersonality": {
                    "type": "string",
                    "enum": [
                      "friendly",
                      "professional",
                      "sarcastic"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Smart Bot settings updated successfully"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  }
}